examples/job-market-analyzer/requirements.txt
# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# LLM
openai>=1.0.0

# Data validation
pydantic>=2.0.0

# Database
# sqlite3 is built-in

# Analytics & Charts
matplotlib>=3.7.0

# Telegram bot
pyTelegramBotAPI>=4.14.0

# Environment
python-dotenv>=1.0.0

# Async
asyncio-throttle>=1.0.0

# Anti-detection (your library)
# phantom-persona>=0.1.0

# Utils
pathlib2>=2.3.0;python_version<"3.4"
